app/backend/requirements.txt
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
fastapi==0.115.6
uvicorn==0.22.0
pillow==10.0.0
numpy==1.25.2
scikit-learn==1.3.0
pandas==2.0.3
albumentations==1.3.1
python-multipart==0.0.6